feature: add support for composite interfaces
This work is based on the supposed capability of reflect to
use methods in embedded fields created by reflect.StructOf.

Unfortunately, this capability comes with many limitations and
caveats, one being that it works only with the embedded type is
the first field, thus forbidding composing.

I will not pursue this experiment further, and consider it as
failed.