feature: generic interface wrapper
This is an experiment of using a generic interface wrapper to implement
composite interfaces.

A generic wrapper is a single type which implements all of the interface
methods provided by the standard library. There is approximately 200
such methods.

This approach, despite being quite simple and allowing to implement
interpreter objects which satisfy several orthogonal interfaces, has
some important drawbacks:

1. storage cost: 200 methods attached to every wrapper. Only a couple
   of them are really active.

2. performance cost:  type assertions has to cope now with this large
   number of methods.

3. behavior change of type assertion. Even if an interpreter object
   doesn't implement an interface, the runtime will detect it does,
   through type assertions. This is problematic, because it changes
   the logic in the runtime, and let the runtime believe that the
   object implements an interface, which is false.

For those reasons, I think this solution should not be selected.
I keep this branch for reference only.